Move stringify functions from tie_db to packet_common
